Fix and JUnits to Issue 824 (Wrong Hop-By-Hop Identifier in answers when > Integer.MAX_VALUE).
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/diameter@7041 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ core/jdiameter/impl/src/main/java/org/jdiameter/client/api/controller/IPeer.java
@@ -33,7 +33,7 @@ public interface IPeer extends Peer {
      * Return new hop by hop id for new message
      * @return new hop by hop id
      */
-    int getHopByHopIdentifier();
+    long getHopByHopIdentifier();
 
     /**
      * Append request to peer request storage map
